# internal/api/static/openapi/openapi.json
{
...
    "info": {
        "description": "HTTP gateway that runs a sequential pipeline: (1) fetch the kube-state-graph response with the inbound query, (2) extract `data.ipaddress` from every `node`-type entry, (3) if any IPs were collected, fetch the switch backend with those IPs batched as `?ip=…`, (4) re-anchor switch shadow nodes onto kube node IDs via IP match, (5) merge into a single Cytoscape.js envelope. Propagates W3C trace context end-to-end and emits structured logs via slog. Any backend failure returns `502`.",
        "title": "graph-api-gateway",
        "version": "v1"
    },
    "openapi": "3.1.0",
    "paths": {
...
        "/docs/assets/{path}": {
            "get": {
                "parameters": [
                    {
                        "description": "Asset path",
                        "in": "path",
                        "name": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "file"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "asset not found"
                    }
                },
                "summary": "API reference asset (vendored)",
                "tags": [
                    "docs"
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}